+++ plugins/repository-azure/src/main/java/org/elasticsearch/repositories/azure/AzureOutputStream.java
@@ -17,14 +17,30 @@
  * under the License.
  */
 
-package org.elasticsearch.cloud.azure;
+package org.elasticsearch.repositories.azure;
 
-public class AzureServiceRemoteException extends IllegalStateException {
-    public AzureServiceRemoteException(String msg) {
-        super(msg);
+import java.io.IOException;
+import java.io.OutputStream;
+
+public class AzureOutputStream extends OutputStream {
+
+    private final OutputStream blobOutputStream;
+
+    public AzureOutputStream(OutputStream blobOutputStream) {
+        this.blobOutputStream = blobOutputStream;
     }
 
-    public AzureServiceRemoteException(String msg, Throwable cause) {
-        super(msg, cause);
+    @Override
+    public void write(int b) throws IOException {
+        blobOutputStream.write(b);
+    }
+
+    @Override
+    public void close() throws IOException {
+        try {
+            blobOutputStream.close();
+        } catch (IOException e) {
+            // Azure is sending a "java.io.IOException: Stream is already closed."
+        }
     }
 }
